a lot of things and working on auton

+++ Gen6/include/custom/systems/lift.hpp
@@ -13,7 +13,7 @@ namespace lift{
   void init(void);
   void pulse(void);
   namespace auton{
-    void lift(double pos, double velocity);
+    void lift(int pos, int velocity);
   }
 }
